refactor(core)!: split engine domains
