+++ packages/core/src/engine.test.ts
@@ -1,10 +1,6 @@
 import { describe, it, expect, beforeEach, afterEach, vi } from 'vite-plus/test';
 import { TimelineEngine } from '#core/engine';
-import {
-  createTimelineScalarKeyframeProperty,
-  getTimelineKeyframeBezierProgress,
-  getTimelineKeyframeBezierControlPoints,
-} from '#core/keyframes';
+import { createTimelineScalarKeyframeProperty } from '#core/keyframes';
 import type { Clip, Track } from '#core/types';
 import type {
   ClipCreatedEvent,
@@ -12,11 +8,7 @@ import type {
   ClipRemovedEvent,
   ClipSplitEvent,
 } from '#core/events';
-import {
-  assertValidRationalTime,
-  fromSeconds,
-  toSeconds,
-} from '@techsquidtv/canvas-timeline-utils';
+import { fromSeconds, toSeconds } from '@techsquidtv/canvas-timeline-utils';
 import { expectDefined } from '#test-utils/assertions';
 
 const opacityKeyframeProperty = createTimelineScalarKeyframeProperty({
@@ -28,18 +20,6 @@ const opacityKeyframeProperty = createTimelineScalarKeyframeProperty({
   formatValue: (value) => `${Math.round(value * 100)}%`,
   getBaseValue: (clip) => clip.opacity ?? 1,
 });
-const nonlinearKeyframeProperty = {
-  id: 'gamma',
-  label: 'Gamma',
-  min: 0,
-  max: 100,
-  defaultValue: 0,
-  clampValue: (value: number) => Math.max(0, Math.min(100, value)),
-  normalizeValue: (value: number) => Math.sqrt(Math.max(0, Math.min(100, value)) / 100),
-  denormalizeValue: (normalized: number) => Math.max(0, Math.min(1, normalized)) ** 2 * 100,
-  formatValue: (value: number) => `${Math.round(value)}`,
-};
-
 type TimelineEngineInternals = {
   historyManager: {
     history: { tracks: string; markers: string; clipGroups: string }[];
@@ -201,522 +181,6 @@ describe('TimelineEngine', () => {
       expect(latestSnapshot.tracks.length).toBeLessThan(2000);
       expect(internals.dragSnapshot).not.toContain(bulky);
       expect(JSON.stringify(internals.clipboardManager.clipboard)).not.toContain(bulky);
-    });
-  });
-
-  describe('Edit command layer', () => {
-    function createEditClip(id: string, start: number, end: number): Clip {
-      return {
-        id,
-        sourceId: `${id}-source`,
-        timelineStart: fromSeconds(start),
-        timelineEnd: fromSeconds(end),
-        sourceStart: fromSeconds(0),
-        selected: false,
-      };
-    }
-
-    function createEditTrack(id: string, clips: Clip[], locked = false): Track {
-      return {
-        id,
-        kind: 'visual',
-        clips,
-        selected: false,
-        locked,
-        muted: false,
-        visible: true,
-      };
-    }
-
-    it('previews commands without mutating state and clears previews on cancel', () => {
-      const preview = engine.previewEdit({
-        type: 'trim',
-        clipId: 'clip1',
-        edge: 'end',
-        newTime: fromSeconds(3),
-        snap: false,
-      });
-
-      expect(preview.valid).toBe(true);
-      expect(preview.changedClips[0].id).toBe('clip1');
-      expect(toSeconds(preview.changedClips[0].timelineEnd)).toBe(3);
-      expect(toSeconds(engine.getClip('clip1')?.clip.timelineEnd ?? fromSeconds(0))).toBe(5);
-      expect(engine.getEditPreview()).toBe(preview);
-      expect(engine.getEditImpacts()?.operation).toBe('trim');
-
-      engine.cancelEdit();
-
-      expect(engine.getEditPreview()).toBeNull();
-      expect(engine.getEditImpacts()).toBeNull();
-    });
-
-    it('commits move commands as one undoable history entry', () => {
-      const moveEvents: ClipMoveEvent[] = [];
-      engine.on('clip:move', (event) => moveEvents.push(event));
-
-      const result = engine.commitEdit({
-        type: 'move',
-        clipId: 'clip1',
-        startTime: fromSeconds(2),
-        snap: false,
-      });
-
-      expect(result.committed).toBe(true);
-      expect(toSeconds(engine.getClip('clip1')?.clip.timelineStart ?? fromSeconds(0))).toBe(2);
-      expect(moveEvents).toHaveLength(1);
-      expect(moveEvents[0].phase).toBe('commit');
-      expect(toSeconds(moveEvents[0].startTime)).toBe(2);
-
-      engine.undo();
-
-      expect(toSeconds(engine.getClip('clip1')?.clip.timelineStart ?? fromSeconds(0))).toBe(1);
-    });
-
-    it('rejects invalid commands before policy and app policy can reject valid commands', () => {
-      expect(
-        engine.validateEdit({
-          type: 'move',
-          clipId: 'missing',
-          startTime: fromSeconds(2),
-        }).reason
-      ).toBe('not-found');
-
-      engine.setEditPolicy({
-        canPlaceClip: () => ({ valid: false, reason: 'policy-rejected' }),
-      });
-
-      expect(
-        engine.commitEdit({
-          type: 'move',
-          clipId: 'clip1',
-          startTime: fromSeconds(2),
-        }).preview.reason
-      ).toBe('policy-rejected');
-      expect(toSeconds(engine.getClip('clip1')?.clip.timelineStart ?? fromSeconds(0))).toBe(1);
-    });
-
-    it('commits insert, overwrite, and range removal commands', () => {
-      const commandEngine = new TimelineEngine({
-        tracks: [
-          createEditTrack('track1', [
-            createEditClip('a', 0, 2),
-            createEditClip('b', 4, 6),
-            createEditClip('c', 8, 10),
-          ]),
-        ],
-      });
-
-      commandEngine.commitEdit({
-        type: 'insert',
-        clip: createEditClip('inserted', 0, 1),
-        targetTrackId: 'track1',
-        startTime: fromSeconds(4),
-        snap: false,
-      });
-
-      expect(
-        commandEngine.tracks[0].clips.map((clip) => [clip.id, toSeconds(clip.timelineStart)])
-      ).toEqual([
-        ['a', 0],
-        ['inserted', 4],
-        ['b', 5],
-        ['c', 9],
-      ]);
-
-      const overwriteResult = commandEngine.commitEdit({
-        type: 'overwrite',
-        clip: createEditClip('winner', 0, 3),
-        targetTrackId: 'track1',
-        startTime: fromSeconds(4.5),
-        snap: false,
-      });
-
-      expect(overwriteResult.preview.removedClips.map((clip) => clip.id)).toContain('b');
-      expect(commandEngine.getClip('winner')).toBeDefined();
-
-      const overwritePreviewEngine = new TimelineEngine({
-        tracks: [createEditTrack('track1', [createEditClip('victim', 0, 10)])],
-      });
-
-      overwritePreviewEngine.previewEdit({
-        type: 'overwrite',
-        clip: createEditClip('preview-winner', 0, 2),
-        targetTrackId: 'track1',
-        startTime: fromSeconds(3),
-        snap: false,
-      });
-
-      expect(overwritePreviewEngine.getEditImpacts()).toMatchObject({
-        operation: 'overwrite',
-        sourceClipId: 'preview-winner',
-        sourceTrackId: 'track1',
-      });
-
-      commandEngine.commitEdit({
-        type: 'delete-range',
-        startTime: fromSeconds(4),
-        endTime: fromSeconds(6),
-        trackIds: ['track1'],
-      });
-
-      const remainingWinner = expectDefined(commandEngine.getClip('winner'), 'winner clip').clip;
-      expect(toSeconds(remainingWinner.timelineStart)).toBe(4);
-      expect(toSeconds(remainingWinner.timelineEnd)).toBe(5.5);
-      expect(
-        commandEngine.tracks[0].clips.every((clip) => toSeconds(clip.timelineStart) <= 7)
-      ).toBe(true);
-    });
-
-    it('previews and commits delete-clips through the command layer', () => {
-      const deleteEngine = new TimelineEngine({
-        clipGroups: [{ id: 'linked-group', clipIds: ['b', 'linked'] }],
-        tracks: [
-          createEditTrack('track1', [createEditClip('a', 0, 2), createEditClip('b', 4, 6)]),
-          createEditTrack('track2', [createEditClip('linked', 4, 6)]),
-        ],
-      });
-      const removed: ClipRemovedEvent[] = [];
-      deleteEngine.on('clip:removed', (event) => removed.push(event));
-
-      const preview = deleteEngine.previewEdit({ type: 'delete-clips', clipIds: ['b'] });
-
-      expect(preview.valid).toBe(true);
-      expect(preview.removedClips.map((clip) => clip.id).sort()).toEqual(['b', 'linked']);
-      expect(deleteEngine.getClip('b')).toBeDefined();
-      expect(deleteEngine.getEditImpacts()).toMatchObject({
-        operation: 'delete-clips',
-        sourceClipId: 'b',
-        sourceTrackId: 'track1',
-      });
-      expect(deleteEngine.getEditImpacts()?.impacts).toHaveLength(2);
-
-      const result = deleteEngine.commitEdit({ type: 'delete-clips', clipIds: ['b'] });
-
-      expect(result.committed).toBe(true);
-      expect(deleteEngine.getClip('b')).toBeUndefined();
-      expect(deleteEngine.getClip('linked')).toBeUndefined();
-      expect(deleteEngine.clipGroups).toEqual([]);
-      expect(removed.map((event) => event.clip.id).sort()).toEqual(['b', 'linked']);
-      expect(removed.every((event) => event.reason === 'delete')).toBe(true);
-
-      deleteEngine.undo();
-
-      expect(deleteEngine.getClip('b')).toBeDefined();
-      expect(deleteEngine.getClip('linked')).toBeDefined();
-      expect(deleteEngine.clipGroups[0]?.clipIds).toEqual(['b', 'linked']);
-    });
-
-    it('rejects delete-clips for locked tracks and policy rejections', () => {
-      const lockedEngine = new TimelineEngine({
-        tracks: [createEditTrack('locked-track', [createEditClip('locked-clip', 0, 2)], true)],
-      });
-
-      expect(
-        lockedEngine.validateEdit({ type: 'delete-clips', clipIds: ['locked-clip'] }).reason
-      ).toBe('locked');
-
-      engine.setEditPolicy({
-        validateCommand: ({ command }) =>
-          command.type === 'delete-clips' ? { valid: false, reason: 'policy-rejected' } : undefined,
-      });
-
-      const result = engine.commitEdit({ type: 'delete-clips', clipIds: ['clip1'] });
-
-      expect(result.committed).toBe(false);
-      expect(result.preview.reason).toBe('policy-rejected');
-      expect(engine.getClip('clip1')).toBeDefined();
-    });
-
-    it('supports ripple trim and roll trim command resolution', () => {
-      const trimEngine = new TimelineEngine({
-        tracks: [
-          createEditTrack('track1', [
-            createEditClip('left', 0, 4),
-            createEditClip('right', 4, 8),
-            createEditClip('later', 10, 12),
-          ]),
-        ],
-      });
-
-      trimEngine.commitEdit({
-        type: 'roll-trim',
-        leftClipId: 'left',
-        rightClipId: 'right',
-        boundaryTime: fromSeconds(5),
-        snap: false,
-      });
-
-      expect(toSeconds(trimEngine.getClip('left')?.clip.timelineEnd ?? fromSeconds(0))).toBe(5);
-      expect(toSeconds(trimEngine.getClip('right')?.clip.timelineStart ?? fromSeconds(0))).toBe(5);
-
-      trimEngine.commitEdit({
-        type: 'ripple-trim',
-        clipId: 'right',
-        edge: 'end',
-        newTime: fromSeconds(7),
-        snap: false,
-      });
-
-      expect(toSeconds(trimEngine.getClip('later')?.clip.timelineStart ?? fromSeconds(0))).toBe(9);
-    });
-
-    it('commits the active preview resolution so split clip ids stay stable', () => {
-      const previewEngine = new TimelineEngine({
-        tracks: [createEditTrack('track1', [createEditClip('victim', 0, 10)])],
-      });
-      const command = {
-        type: 'overwrite' as const,
-        clip: createEditClip('winner', 0, 2),
-        targetTrackId: 'track1',
-        startTime: fromSeconds(3),
-        snap: false,
-      };
-      const preview = previewEngine.previewEdit(command);
-      const previewSplitClip = expectDefined(
-        preview.createdClips.find((clip) => clip.id !== 'winner'),
-        'preview split clip'
-      );
-
-      previewEngine.commitEdit(command);
-
-      expect(previewEngine.getClip(previewSplitClip.id)).toBeDefined();
-    });
-
-    it('commits active grouped overwrite previews so split clip ids stay stable', () => {
-      const previewEngine = new TimelineEngine({
-        tracks: [
-          createEditTrack('track1', [createEditClip('video-victim', 0, 10)]),
-          createEditTrack('track2', []),
-        ],
-      });
-      const command = {
-        type: 'overwrite-clip-group' as const,
-        groupId: 'preview-group',
-        placements: [
-          {
-            clip: createEditClip('video-winner', 0, 2),
-            targetTrackId: 'track1',
-            startTime: fromSeconds(3),
-          },
-          {
-            clip: createEditClip('audio-winner', 0, 2),
-            targetTrackId: 'track2',
-            startTime: fromSeconds(3),
-          },
-        ],
-        snap: false,
-      };
-      const preview = previewEngine.previewEdit(command);
-      const previewSplitClip = expectDefined(
-        preview.createdClips.find(
-          (clip) => clip.id !== 'video-winner' && clip.id !== 'audio-winner'
-        ),
-        'preview split clip'
-      );
-
-      previewEngine.commitEdit(command);
-
-      expect(previewEngine.getClip(previewSplitClip.id)).toBeDefined();
-      expect(previewEngine.getClipGroup('preview-group')?.clipIds).toEqual([
-        'video-winner',
-        'audio-winner',
-      ]);
-    });
-
-    it('rejects grouped placement commands through edit policy without mutation', () => {
-      const policyEngine = new TimelineEngine({
-        tracks: [createEditTrack('track1', []), createEditTrack('track2', [])],
-      });
-      policyEngine.setEditPolicy({
-        canEditRange: (context) =>
-          context.range?.trackId === 'track2'
-            ? { valid: false, reason: 'policy-rejected' }
-            : undefined,
-      });
-      const beforeTracks = JSON.stringify(policyEngine.tracks);
-      const beforeGroups = JSON.stringify(policyEngine.clipGroups);
-
-      const result = policyEngine.commitEdit({
-        type: 'insert-clip-group',
-        groupId: 'policy-group',
-        placements: [
-          {
-            clip: createEditClip('video-drop', 0, 2),
-            targetTrackId: 'track1',
-            startTime: fromSeconds(1),
-          },
-          {
-            clip: createEditClip('audio-drop', 0, 2),
-            targetTrackId: 'track2',
-            startTime: fromSeconds(1),
-          },
-        ],
-        snap: false,
-      });
-
-      expect(result.committed).toBe(false);
-      expect(result.preview.reason).toBe('policy-rejected');
-      expect(JSON.stringify(policyEngine.tracks)).toBe(beforeTracks);
-      expect(JSON.stringify(policyEngine.clipGroups)).toBe(beforeGroups);
-    });
-
-    it('emits lifecycle events and restores clip groups for grouped overwrite undo and redo', () => {
-      const lifecycleEngine = new TimelineEngine({
-        tracks: [
-          createEditTrack('track1', [createEditClip('video-victim', 2, 4)]),
-          createEditTrack('track2', [createEditClip('audio-victim', 2, 4)]),
-        ],
-      });
-      const created: ClipCreatedEvent[] = [];
-      const removed: ClipRemovedEvent[] = [];
-      lifecycleEngine.on('clip:created', (event) => created.push(event));
-      lifecycleEngine.on('clip:removed', (event) => removed.push(event));
-
-      const result = lifecycleEngine.commitEdit({
-        type: 'overwrite-clip-group',
-        groupId: 'lifecycle-group',
-        placements: [
-          {
-            clip: createEditClip('video-drop', 0, 2),
-            targetTrackId: 'track1',
-            startTime: fromSeconds(2),
-          },
-          {
-            clip: createEditClip('audio-drop', 0, 2),
-            targetTrackId: 'track2',
-            startTime: fromSeconds(2),
-          },
-        ],
-        snap: false,
-      });
-
-      expect(result.committed).toBe(true);
-      expect(created.map((event) => [event.reason, event.clip.id])).toEqual([
-        ['overwrite', 'video-drop'],
-        ['overwrite', 'audio-drop'],
-      ]);
-      expect(removed.map((event) => [event.reason, event.clip.id])).toEqual([
-        ['overwrite', 'video-victim'],
-        ['overwrite', 'audio-victim'],
-      ]);
-      expect(lifecycleEngine.getClipGroup('lifecycle-group')?.clipIds).toEqual([
-        'video-drop',
-        'audio-drop',
-      ]);
-
-      lifecycleEngine.undo();
-
-      expect(lifecycleEngine.getClipGroup('lifecycle-group')).toBeUndefined();
-      expect(lifecycleEngine.getClip('video-drop')).toBeUndefined();
-      expect(lifecycleEngine.getClip('audio-drop')).toBeUndefined();
-      expect(lifecycleEngine.getClip('video-victim')).toBeDefined();
-      expect(lifecycleEngine.getClip('audio-victim')).toBeDefined();
-
-      lifecycleEngine.redo();
-
-      expect(lifecycleEngine.getClipGroup('lifecycle-group')?.clipIds).toEqual([
-        'video-drop',
-        'audio-drop',
-      ]);
-      expect(lifecycleEngine.getClip('video-victim')).toBeUndefined();
-      expect(lifecycleEngine.getClip('audio-victim')).toBeUndefined();
-    });
-
-    it('preserves created clip lineage for overwrite and range split commands', () => {
-      const lineageEngine = new TimelineEngine({
-        tracks: [createEditTrack('track1', [createEditClip('victim', 0, 10)])],
-      });
-      const created: ClipCreatedEvent[] = [];
-      lineageEngine.on('clip:created', (event) => created.push(event));
-
-      lineageEngine.commitEdit({
-        type: 'overwrite',
-        clip: createEditClip('winner', 0, 2),
-        targetTrackId: 'track1',
-        startTime: fromSeconds(3),
-        snap: false,
-      });
-
-      expect(
-        created.some((event) => event.reason === 'overwrite' && event.clip.id === 'winner')
-      ).toBe(true);
-      const splitEvent = expectDefined(
-        created.find((event) => event.reason === 'overwrite-split'),
-        'overwrite split event'
-      );
-      expect(splitEvent.originClipId).toBe('victim');
-
-      const rangeEngine = new TimelineEngine({
-        tracks: [createEditTrack('track1', [createEditClip('range-victim', 0, 10)])],
-      });
-      const rangeCreated: ClipCreatedEvent[] = [];
-      rangeEngine.on('clip:created', (event) => rangeCreated.push(event));
-
-      rangeEngine.commitEdit({
-        type: 'lift-range',
-        startTime: fromSeconds(3),
-        endTime: fromSeconds(5),
-        trackIds: ['track1'],
-      });
-
-      const rangeSplitEvent = expectDefined(
-        rangeCreated.find((event) => event.reason === 'range-split'),
-        'range split event'
-      );
-      expect(rangeSplitEvent.originClipId).toBe('range-victim');
-    });
-
-    it('clears snap guides after command commits', () => {
-      const snapEngine = new TimelineEngine({
-        markers: [{ id: 'marker-1', time: fromSeconds(2), label: 'Marker' }],
-        tracks: [createEditTrack('track1', [createEditClip('snap-clip', 0, 1)])],
-      });
-
-      snapEngine.prepareSnapping('snap-clip');
-      snapEngine.previewEdit({
-        type: 'move',
-        clipId: 'snap-clip',
-        startTime: fromSeconds(2.05),
-      });
-
-      expect(snapEngine.getState().snapFeedback.target?.kind).toBe('marker');
-
-      snapEngine.commitEdit({
-        type: 'move',
-        clipId: 'snap-clip',
-        startTime: fromSeconds(2.05),
-      });
-
-      expect(snapEngine.getState().snapFeedback.target).toBeNull();
-      expect(snapEngine.getState().snapFeedback.lines).toEqual([]);
-    });
-
-    it('rejects roll trims when snapping would violate minimum clip duration', () => {
-      const rollEngine = new TimelineEngine({
-        tracks: [
-          createEditTrack('track1', [createEditClip('left', 0, 1), createEditClip('right', 1, 2)]),
-        ],
-        zoomScale: 10,
-      });
-      rollEngine.prepareSnapping();
-
-      const validation = rollEngine.validateEdit({
-        type: 'roll-trim',
-        leftClipId: 'left',
-        rightClipId: 'right',
-        boundaryTime: fromSeconds(0.04),
-      });
-
-      expect(validation).toEqual({ valid: false, reason: 'invalid-duration' });
-      expect(
-        rollEngine.commitEdit({
-          type: 'roll-trim',
-          leftClipId: 'left',
-          rightClipId: 'right',
-          boundaryTime: fromSeconds(0.04),
-        }).committed
-      ).toBe(false);
-      expect(toSeconds(rollEngine.getClip('left')?.clip.timelineEnd ?? fromSeconds(0))).toBe(1);
     });
   });
 
@@ -1881,29 +1345,6 @@ describe('TimelineEngine', () => {
   });
 
   describe('Editing constraints and preview', () => {
-    function createConstraintClip(id: string, start: number, end: number): Clip {
-      return {
-        id,
-        sourceId: `${id}-source`,
-        timelineStart: fromSeconds(start),
-        timelineEnd: fromSeconds(end),
-        sourceStart: fromSeconds(0),
-        selected: false,
-      };
-    }
-
-    function createConstraintTrack(id: string, clips: Clip[], kind = 'visual'): Track {
-      return {
-        id,
-        kind,
-        clips,
-        selected: false,
-        locked: false,
-        muted: false,
-        visible: true,
-      };
-    }
-
     it('should clamp start trims to minStart', () => {
       engine.getState().tracks[0].clips[0].minStart = fromSeconds(2);
 
@@ -2133,602 +1574,6 @@ describe('TimelineEngine', () => {
       expect(toSeconds(payload.right.timelineStart)).toBeCloseTo(2);
     });
 
-    it('groups arbitrary clips, expands selection, and moves linked clips together', () => {
-      const groupedEngine = new TimelineEngine({
-        tracks: [
-          {
-            id: 'video',
-            kind: 'visual',
-            clips: [
-              {
-                id: 'video-clip',
-                sourceId: 'source',
-                timelineStart: fromSeconds(1),
-                timelineEnd: fromSeconds(5),
-                sourceStart: fromSeconds(0),
-                selected: false,
-              },
-            ],
-            selected: false,
-            locked: false,
-            muted: false,
-            visible: true,
-          },
-          {
-            id: 'audio',
-            kind: 'audio',
-            clips: [
-              {
-                id: 'audio-clip',
-                sourceId: 'source',
-                timelineStart: fromSeconds(2),
-                timelineEnd: fromSeconds(6),
-                sourceStart: fromSeconds(0),
-                selected: false,
-              },
-            ],
-            selected: false,
-            locked: false,
-            muted: false,
-            visible: true,
-          },
-        ],
-      });
-
-      const group = expectDefined(
-        groupedEngine.createClipGroup({ id: 'linked-av', clipIds: ['video-clip', 'audio-clip'] }),
-        'created clip group'
-      );
-      expect(group.clipIds).toEqual(['video-clip', 'audio-clip']);
-
-      groupedEngine.selectClip('video-clip');
-      expect(groupedEngine.getClip('video-clip')?.clip.selected).toBe(true);
-      expect(groupedEngine.getClip('audio-clip')?.clip.selected).toBe(true);
-
-      const move = groupedEngine.commitEdit({
-        type: 'move',
-        clipId: 'video-clip',
-        startTime: fromSeconds(4),
-      });
-      expect(move.committed).toBe(true);
-      expect(move.preview.changedClips.map((clip) => clip.id).sort()).toEqual([
-        'audio-clip',
-        'video-clip',
-      ]);
-      expect(
-        toSeconds(groupedEngine.getClip('video-clip')?.clip.timelineStart ?? fromSeconds(0))
-      ).toBeCloseTo(4);
-      expect(
-        toSeconds(groupedEngine.getClip('audio-clip')?.clip.timelineStart ?? fromSeconds(0))
-      ).toBeCloseTo(5);
-    });
-
-    it('rejects duplicate group ids in initial clip group state', () => {
-      expect(
-        () =>
-          new TimelineEngine({
-            tracks: [
-              createConstraintTrack('video', [
-                createConstraintClip('clip-a', 0, 2),
-                createConstraintClip('clip-b', 3, 5),
-                createConstraintClip('clip-c', 6, 8),
-                createConstraintClip('clip-d', 9, 11),
-              ]),
-            ],
-            clipGroups: [
-              { id: 'duplicate-group', clipIds: ['clip-a', 'clip-b'] },
-              { id: 'duplicate-group', clipIds: ['clip-c', 'clip-d'] },
-            ],
-          })
-      ).toThrow('duplicate clip group id "duplicate-group".');
-    });
-
-    it('rejects clips assigned to more than one initial clip group', () => {
-      expect(
-        () =>
-          new TimelineEngine({
-            tracks: [
-              createConstraintTrack('video', [
-                createConstraintClip('clip-a', 0, 2),
-                createConstraintClip('clip-b', 3, 5),
-                createConstraintClip('clip-c', 6, 8),
-              ]),
-            ],
-            clipGroups: [
-              { id: 'group-a', clipIds: ['clip-a', 'clip-b'] },
-              { id: 'group-b', clipIds: ['clip-b', 'clip-c'] },
-            ],
-          })
-      ).toThrow('clip "clip-b" belongs to more than one clip group.');
-    });
-
-    it('inserts clip groups atomically and shifts placed keyframes', () => {
-      const groupedEngine = new TimelineEngine({
-        tracks: [createConstraintTrack('video', []), createConstraintTrack('audio', [], 'audio')],
-        keyframeProperties: [opacityKeyframeProperty],
-      });
-      const group = groupedEngine.insertClipGroup({
-        groupId: 'import-group',
-        placements: [
-          {
-            clip: {
-              ...createConstraintClip('video-clip', 0, 2),
-              keyframes: [
-                { id: 'video-opacity', property: 'opacity', time: fromSeconds(1), value: 1 },
-              ],
-            },
-            targetTrackId: 'video',
-            startTime: fromSeconds(5),
-          },
-          {
-            clip: createConstraintClip('audio-clip', 0, 2),
-            targetTrackId: 'audio',
-            startTime: fromSeconds(5),
-          },
-        ],
-      });
-
-      expect(group?.clipIds).toEqual(['video-clip', 'audio-clip']);
-      expect(
-        toSeconds(groupedEngine.getClip('video-clip')?.clip.timelineStart ?? fromSeconds(0))
-      ).toBe(5);
-      expect(
-        groupedEngine
-          .getClip('video-clip')
-          ?.clip.keyframes?.map((keyframe) => toSeconds(keyframe.time))
-      ).toEqual([6]);
-
-      const failed = groupedEngine.insertClipGroup({
-        placements: [
-          {
-            clip: createConstraintClip('valid-clip', 0, 2),
-            targetTrackId: 'video',
-            startTime: fromSeconds(8),
-          },
-          {
-            clip: {
-              ...createConstraintClip('invalid-clip', 0, 2),
-              timelineStart: { v: 0.5, r: 60000 },
-            },
-            targetTrackId: 'audio',
-            startTime: fromSeconds(8),
-          },
-        ],
-      });
-
-      expect(failed).toBeNull();
-      expect(groupedEngine.getClip('valid-clip')).toBeUndefined();
-      expect(groupedEngine.getClip('invalid-clip')).toBeUndefined();
-      expect(groupedEngine.clipGroups).toHaveLength(1);
-    });
-
-    it('commits grouped insert commands atomically and ripples affected target tracks', () => {
-      const groupedEngine = new TimelineEngine({
-        tracks: [
-          createConstraintTrack('video', [createConstraintClip('video-after', 5, 7)]),
-          createConstraintTrack('audio', [createConstraintClip('audio-after', 6, 8)], 'audio'),
-        ],
-      });
-      groupedEngine.setSnappingEnabled(false);
-
-      const result = groupedEngine.commitEdit({
-        type: 'insert-clip-group',
-        groupId: 'dropped-av',
-        label: 'Dropped AV',
-        placements: [
-          {
-            clip: createConstraintClip('video-drop', 0, 2),
-            targetTrackId: 'video',
-            startTime: fromSeconds(3),
-          },
-          {
-            clip: createConstraintClip('audio-drop', 0, 3),
-            targetTrackId: 'audio',
-            startTime: fromSeconds(4),
-          },
-        ],
-        snap: false,
-      });
-
-      expect(result.committed).toBe(true);
-      expect(result.preview.createdClips.map((clip) => clip.id)).toEqual([
-        'video-drop',
-        'audio-drop',
-      ]);
-      expect(result.preview.changedClips.map((clip) => clip.id).sort()).toEqual([
-        'audio-after',
-        'video-after',
-      ]);
-      expect(groupedEngine.getClipGroup('dropped-av')).toMatchObject({
-        id: 'dropped-av',
-        label: 'Dropped AV',
-        clipIds: ['video-drop', 'audio-drop'],
-      });
-      expect(
-        toSeconds(groupedEngine.getClip('video-after')?.clip.timelineStart ?? fromSeconds(0))
-      ).toBe(7);
-      expect(
-        toSeconds(groupedEngine.getClip('audio-after')?.clip.timelineStart ?? fromSeconds(0))
-      ).toBe(9);
-    });
-
-    it('snaps grouped insert commands with one shared delta to preserve relative offsets', () => {
-      const groupedEngine = new TimelineEngine({
-        markers: [
-          { id: 'primary-marker', time: fromSeconds(3), label: 'Primary' },
-          { id: 'secondary-marker', time: fromSeconds(6.9), label: 'Secondary' },
-        ],
-        tracks: [createConstraintTrack('video', []), createConstraintTrack('audio', [], 'audio')],
-        zoomScale: 100,
-      });
-      groupedEngine.prepareSnapping();
-
-      const result = groupedEngine.commitEdit({
-        type: 'insert-clip-group',
-        groupId: 'snapped-av',
-        placements: [
-          {
-            clip: createConstraintClip('video-drop', 0, 2),
-            targetTrackId: 'video',
-            startTime: fromSeconds(3.05),
-          },
-          {
-            clip: createConstraintClip('audio-drop', 0, 2),
-            targetTrackId: 'audio',
-            startTime: fromSeconds(6.86),
-          },
-        ],
-      });
-
-      expect(result.committed).toBe(true);
-      expect(
-        toSeconds(groupedEngine.getClip('video-drop')?.clip.timelineStart ?? fromSeconds(0))
-      ).toBeCloseTo(3);
-      expect(
-        toSeconds(groupedEngine.getClip('audio-drop')?.clip.timelineStart ?? fromSeconds(0))
-      ).toBeCloseTo(6.81);
-    });
-
-    it('commits grouped overwrite commands atomically and trims overlaps per target track', () => {
-      const groupedEngine = new TimelineEngine({
-        tracks: [
-          createConstraintTrack('video', [createConstraintClip('video-victim', 1, 6)]),
-          createConstraintTrack('audio', [createConstraintClip('audio-victim', 2, 8)], 'audio'),
-        ],
-      });
-      groupedEngine.setSnappingEnabled(false);
-
-      const result = groupedEngine.commitEdit({
-        type: 'overwrite-clip-group',
-        groupId: 'overwrite-av',
-        placements: [
-          {
-            clip: createConstraintClip('video-drop', 0, 2),
-            targetTrackId: 'video',
-            startTime: fromSeconds(3),
-          },
-          {
-            clip: createConstraintClip('audio-drop', 0, 3),
-            targetTrackId: 'audio',
-            startTime: fromSeconds(4),
-          },
-        ],
-        snap: false,
-      });
-
-      expect(result.committed).toBe(true);
-      expect(groupedEngine.getClipGroup('overwrite-av')?.clipIds).toEqual([
-        'video-drop',
-        'audio-drop',
-      ]);
-      expect(result.preview.impacts.map((impact) => impact.clipId).sort()).toEqual([
-        'audio-victim',
-        'video-victim',
-      ]);
-      expect(groupedEngine.getClip('video-victim')).toBeDefined();
-      expect(groupedEngine.getClip('audio-victim')).toBeDefined();
-      expect(groupedEngine.tracks[0].clips.map((clip) => clip.id)).toHaveLength(3);
-      expect(groupedEngine.tracks[1].clips.map((clip) => clip.id)).toHaveLength(3);
-    });
-
-    it('rejects invalid grouped placement commands without mutating tracks or groups', () => {
-      const groupedEngine = new TimelineEngine({
-        tracks: [createConstraintTrack('video', [createConstraintClip('existing', 0, 2)])],
-      });
-      const beforeTracks = JSON.stringify(groupedEngine.tracks);
-      const beforeGroups = JSON.stringify(groupedEngine.clipGroups);
-
-      const result = groupedEngine.commitEdit({
-        type: 'insert-clip-group',
-        placements: [
-          {
-            clip: createConstraintClip('new-a', 0, 2),
-            targetTrackId: 'video',
-            startTime: fromSeconds(3),
-          },
-          {
-            clip: createConstraintClip('new-a', 0, 2),
-            targetTrackId: 'missing',
-            startTime: fromSeconds(3),
-          },
-        ],
-        snap: false,
-      });
-
-      expect(result.committed).toBe(false);
-      expect(result.preview.reason).toBe('duplicate-id');
-      expect(JSON.stringify(groupedEngine.tracks)).toBe(beforeTracks);
-      expect(JSON.stringify(groupedEngine.clipGroups)).toBe(beforeGroups);
-    });
-
-    it('applies overwrite cleanup for every grouped member during drag preview', () => {
-      const groupedEngine = new TimelineEngine({
-        tracks: [
-          createConstraintTrack('video', [createConstraintClip('video-clip', 0, 2)]),
-          createConstraintTrack(
-            'audio',
-            [createConstraintClip('audio-clip', 0, 2), createConstraintClip('audio-victim', 4, 6)],
-            'audio'
-          ),
-        ],
-      });
-      groupedEngine.createClipGroup({ id: 'linked-av', clipIds: ['video-clip', 'audio-clip'] });
-
-      groupedEngine.startDrag();
-      expect(
-        groupedEngine.moveClip({
-          clipId: 'video-clip',
-          startTime: fromSeconds(4),
-          snap: false,
-        })
-      ).toBe(true);
-
-      expect(groupedEngine.getClip('audio-victim')).toBeUndefined();
-      groupedEngine.endDrag();
-    });
-
-    it('splits selected grouped clips and repartitions the group at the blade time', () => {
-      const groupedEngine = new TimelineEngine({
-        tracks: [
-          {
-            id: 'video',
-            kind: 'visual',
-            clips: [
-              {
-                id: 'video-clip',
-                sourceId: 'source',
-                timelineStart: fromSeconds(0),
-                timelineEnd: fromSeconds(10),
-                sourceStart: fromSeconds(0),
-                selected: false,
-              },
-            ],
-            selected: false,
-            locked: false,
-            muted: false,
-            visible: true,
-          },
-          {
-            id: 'audio',
-            kind: 'audio',
-            clips: [
-              {
-                id: 'audio-clip',
-                sourceId: 'source',
-                timelineStart: fromSeconds(0),
-                timelineEnd: fromSeconds(10),
-                sourceStart: fromSeconds(0),
-                selected: false,
-              },
-            ],
-            selected: false,
-            locked: false,
-            muted: false,
-            visible: true,
-          },
-        ],
-      });
-      groupedEngine.createClipGroup({ id: 'linked-av', clipIds: ['video-clip', 'audio-clip'] });
-      groupedEngine.selectClip('video-clip');
-
-      const split = groupedEngine.commitEdit({
-        type: 'split',
-        clipIds: ['video-clip', 'audio-clip'],
-        time: fromSeconds(4),
-      });
-
-      expect(split.committed).toBe(true);
-      expect(split.preview.createdClips).toHaveLength(2);
-      expect(groupedEngine.clipGroups).toHaveLength(2);
-      expect(groupedEngine.getClipGroup('linked-av')?.clipIds).toEqual([
-        'video-clip',
-        'audio-clip',
-      ]);
-      const rightGroup = expectDefined(
-        groupedEngine.clipGroups.find((group) => group.id !== 'linked-av'),
-        'right-side group'
-      );
-      expect(rightGroup.clipIds).toHaveLength(2);
-      expect(
-        rightGroup.clipIds.every((clipId) => groupedEngine.getClip(clipId) !== undefined)
-      ).toBe(true);
-    });
-
-    it('keeps grouped split pieces valid across repeated mixed-rate moves', () => {
-      const groupedEngine = new TimelineEngine({
-        tracks: [
-          {
-            id: 'video',
-            kind: 'visual',
-            clips: [
-              {
-                id: 'video-clip',
-                sourceId: 'source',
-                timelineStart: fromSeconds(3),
-                timelineEnd: fromSeconds(9),
-                sourceStart: fromSeconds(0),
-                selected: false,
-              },
-            ],
-            selected: false,
-            locked: false,
-            muted: false,
-            visible: true,
-          },
-          {
-            id: 'audio',
-            kind: 'audio',
-            clips: [
-              {
-                id: 'audio-clip',
-                sourceId: 'source',
-                timelineStart: fromSeconds(3),
-                timelineEnd: fromSeconds(9),
-                sourceStart: fromSeconds(0),
-                selected: false,
-              },
-            ],
-            selected: false,
-            locked: false,
-            muted: false,
-            visible: true,
-          },
-        ],
-      });
-      groupedEngine.createClipGroup({ id: 'linked-av', clipIds: ['video-clip', 'audio-clip'] });
-      groupedEngine.selectClip('video-clip');
-      groupedEngine.commitEdit({
-        type: 'split',
-        clipIds: ['video-clip', 'audio-clip'],
-        time: fromSeconds(6.5),
-      });
-
-      for (let index = 0; index < 8; index += 1) {
-        expect(
-          groupedEngine.moveClip({
-            clipId: 'video-clip',
-            startTime: fromSeconds(4 + index * 0.01, 24000),
-          })
-        ).toBe(true);
-      }
-
-      const videoClip = expectDefined(groupedEngine.getClip('video-clip')?.clip, 'video clip');
-      const audioClip = expectDefined(groupedEngine.getClip('audio-clip')?.clip, 'audio clip');
-      assertValidRationalTime(videoClip.timelineStart, 'videoClip.timelineStart');
-      assertValidRationalTime(audioClip.timelineStart, 'audioClip.timelineStart');
-      expect(videoClip.timelineStart.r).toBeLessThanOrEqual(120000);
-      expect(audioClip.timelineStart.r).toBeLessThanOrEqual(120000);
-      expect(toSeconds(audioClip.timelineStart)).toBeCloseTo(toSeconds(videoClip.timelineStart));
-    });
-
-    it('rejects grouped splits when an overlapping linked member cannot be split', () => {
-      const groupedEngine = new TimelineEngine({
-        tracks: [
-          createConstraintTrack('video', [createConstraintClip('video-clip', 0, 10)]),
-          createConstraintTrack(
-            'audio',
-            [
-              {
-                ...createConstraintClip('audio-clip', 0, 10),
-                resizable: false,
-              },
-            ],
-            'audio'
-          ),
-        ],
-      });
-      groupedEngine.createClipGroup({ id: 'linked-av', clipIds: ['video-clip', 'audio-clip'] });
-
-      const split = groupedEngine.commitEdit({
-        type: 'split',
-        clipIds: ['video-clip'],
-        time: fromSeconds(4),
-      });
-
-      expect(split.committed).toBe(false);
-      expect(split.preview.reason).toBe('locked');
-      expect(groupedEngine.getClip('video-clip')?.track.clips).toHaveLength(1);
-      expect(groupedEngine.getClip('audio-clip')?.track.clips).toHaveLength(1);
-      expect(groupedEngine.getClipGroup('linked-av')?.clipIds).toEqual([
-        'video-clip',
-        'audio-clip',
-      ]);
-    });
-
-    it('skips non-overlapping selected clips when splitting at a blade time', () => {
-      const splitEngine = new TimelineEngine({
-        tracks: [
-          createConstraintTrack('track1', [
-            createConstraintClip('selected-overlap', 0, 10),
-            {
-              ...createConstraintClip('selected-later', 20, 30),
-              resizable: false,
-            },
-          ]),
-        ],
-      });
-
-      const split = splitEngine.commitEdit({
-        type: 'split',
-        clipIds: ['selected-overlap', 'selected-later'],
-        time: fromSeconds(4),
-      });
-
-      expect(split.committed).toBe(true);
-      expect(split.preview.createdClips).toHaveLength(1);
-      expect(splitEngine.getClip('selected-overlap')?.track.clips).toHaveLength(3);
-    });
-
-    it('preserves copied group membership on paste and restores groups through history', () => {
-      const groupedEngine = new TimelineEngine({
-        tracks: [
-          {
-            id: 'track1',
-            kind: 'visual',
-            clips: [
-              {
-                id: 'clip-a',
-                sourceId: 'source-a',
-                timelineStart: fromSeconds(0),
-                timelineEnd: fromSeconds(2),
-                sourceStart: fromSeconds(0),
-                selected: false,
-              },
-              {
-                id: 'clip-b',
-                sourceId: 'source-b',
-                timelineStart: fromSeconds(3),
-                timelineEnd: fromSeconds(5),
-                sourceStart: fromSeconds(0),
-                selected: false,
-              },
-            ],
-            selected: false,
-            locked: false,
-            muted: false,
-            visible: true,
-          },
-        ],
-      });
-
-      groupedEngine.createClipGroup({ id: 'copy-group', clipIds: ['clip-a', 'clip-b'] });
-      groupedEngine.undo();
-      expect(groupedEngine.clipGroups).toEqual([]);
-      groupedEngine.redo();
-      expect(groupedEngine.getClipGroup('copy-group')?.clipIds).toEqual(['clip-a', 'clip-b']);
-
-      groupedEngine.selectClip('clip-a');
-      groupedEngine.copySelection();
-      groupedEngine.pasteSelection(fromSeconds(10));
-
-      expect(groupedEngine.clipGroups).toHaveLength(2);
-      const pastedGroup = expectDefined(
-        groupedEngine.clipGroups.find((group) => group.id !== 'copy-group'),
-        'pasted group'
-      );
-      expect(pastedGroup.clipIds).toHaveLength(2);
-    });
-
     it('emits clip lifecycle events for committed overwrite edits', () => {
       const overwriteEngine = new TimelineEngine({
         tracks: [
@@ -2914,882 +1759,6 @@ describe('TimelineEngine', () => {
     it('should set track group', () => {
       engine.setTrackGroup('track1', 'groupA');
       expect(engine.getState().tracks[0].groupId).toBe('groupA');
-    });
-  });
-
-  describe('Keyframes', () => {
-    function createKeyframeClip(id: string, start: number, end: number, values: number[]): Clip {
-      return {
-        id,
-        sourceId: `${id}-source`,
-        timelineStart: fromSeconds(start),
-        timelineEnd: fromSeconds(end),
-        sourceStart: fromSeconds(0),
-        selected: false,
-        keyframes: values.map((time, index) => ({
-          id: `${id}-opacity-${index}`,
-          property: 'opacity',
-          time: fromSeconds(time),
-          value: index / Math.max(1, values.length - 1),
-        })),
-      };
-    }
-
-    function createKeyframeTrack(clips: Clip[]): Track {
-      return {
-        id: 'keyframe-track',
-        kind: 'visual',
-        clips,
-        selected: false,
-        locked: false,
-        muted: false,
-        visible: true,
-      };
-    }
-
-    function keyframeSeconds(clip: Clip | undefined): number[] {
-      return (clip?.keyframes ?? []).map((keyframe) => toSeconds(keyframe.time));
-    }
-
-    it('stores registered keyframe properties as immutable snapshots', () => {
-      const registryEngine = new TimelineEngine({ tracks: [] });
-      const scaleOptions = {
-        id: 'scale',
-        label: 'Scale',
-        min: 0,
-        max: 4,
-        defaultValue: 1,
-      };
-      const scaleProperty = createTimelineScalarKeyframeProperty(scaleOptions);
-
-      registryEngine.registerKeyframeProperty(scaleProperty);
-      scaleOptions.max = 12;
-      scaleProperty.label = 'Changed';
-
-      const registered = expectDefined(
-        registryEngine.getKeyframePropertyDefinition('scale'),
-        'registered keyframe property'
-      );
-      expect(Object.isFrozen(registered)).toBe(true);
-      expect(registered.label).toBe('Scale');
-      expect(registered.clampValue(8)).toBe(4);
-      expect(registered.normalizeValue(2)).toBe(0.5);
-      expect(registryEngine.hasKeyframeProperty('scale')).toBe(true);
-      expect(registryEngine.listKeyframeProperties()).toEqual([registered]);
-    });
-
-    it('rejects keyframe property definitions with invalid mapping outputs', () => {
-      expect(
-        () =>
-          new TimelineEngine({
-            tracks: [],
-            keyframeProperties: [
-              {
-                ...nonlinearKeyframeProperty,
-                id: 'bad-clamp',
-                clampValue: () => Number.NaN,
-              },
-            ],
-          })
-      ).toThrow('clampValue result');
-      expect(
-        () =>
-          new TimelineEngine({
-            tracks: [],
-            keyframeProperties: [
-              {
-                ...nonlinearKeyframeProperty,
-                id: 'bad-normalize',
-                normalizeValue: () => 2,
-              },
-            ],
-          })
-      ).toThrow('normalizeValue must return a value between 0 and 1');
-      expect(
-        () =>
-          new TimelineEngine({
-            tracks: [],
-            keyframeProperties: [
-              {
-                ...nonlinearKeyframeProperty,
-                id: 'bad-denormalize',
-                denormalizeValue: () => 101,
-              },
-            ],
-          })
-      ).toThrow('denormalizeValue must return a value within min/max');
-    });
-
-    it('rejects invalid custom property outputs during evaluation', () => {
-      const unstableKeyframeProperty = {
-        id: 'unstable',
-        label: 'Unstable',
-        min: 0,
-        max: 1,
-        defaultValue: 0,
-        clampValue: (value: number) => Math.max(0, Math.min(1, value)),
-        normalizeValue: (value: number) => Math.max(0, Math.min(1, value)),
-        denormalizeValue: (normalized: number) =>
-          normalized > 0.49 && normalized < 0.51 ? Number.NaN : normalized,
-      };
-      const unstableEngine = new TimelineEngine({
-        tracks: [
-          {
-            ...mockTrack,
-            clips: [
-              {
-                ...mockClip,
-                timelineStart: fromSeconds(0),
-                timelineEnd: fromSeconds(10),
-                keyframes: [
-                  {
-                    id: 'unstable-start',
-                    property: 'unstable',
-                    time: fromSeconds(0),
-                    value: 0,
-                  },
-                  {
-                    id: 'unstable-end',
-                    property: 'unstable',
-                    time: fromSeconds(10),
-                    value: 1,
-                  },
-                ],
-              },
-            ],
-          },
-        ],
-        keyframeProperties: [unstableKeyframeProperty],
-      });
-
-      expect(() =>
-        unstableEngine.getClipPropertyValueAtTime('clip1', 'unstable', fromSeconds(5))
-      ).toThrow('denormalizeValue result');
-    });
-
-    it('evaluates nonlinear properties in normalized space and prepares matching geometry', () => {
-      const nonlinearEngine = new TimelineEngine({
-        tracks: [
-          {
-            ...mockTrack,
-            clips: [
-              {
-                ...mockClip,
-                timelineStart: fromSeconds(0),
-                timelineEnd: fromSeconds(10),
-                keyframes: [
-                  {
-                    id: 'gamma-start',
-                    property: 'gamma',
-                    time: fromSeconds(0),
-                    value: 0,
-                  },
-                  {
-                    id: 'gamma-middle',
-                    property: 'gamma',
-                    time: fromSeconds(5),
-                    value: 25,
-                  },
-                  {
-                    id: 'gamma-end',
-                    property: 'gamma',
-                    time: fromSeconds(10),
-                    value: 100,
-                  },
-                ],
-              },
-            ],
-          },
-        ],
-        keyframeProperties: [nonlinearKeyframeProperty],
-        zoomScale: 100,
-      });
-
-      expect(
-        nonlinearEngine.getClipPropertyValueAtTime('clip1', 'gamma', fromSeconds(2.5))
-      ).toBeCloseTo(6.25);
-
-      const geometry = nonlinearEngine.getKeyframeRenderGeometry({
-        property: 'gamma',
-        rulerHeight: 32,
-        trackHeight: 48,
-        viewportHeight: 100,
-        viewportWidth: 1000,
-      });
-      const renderClip = expectDefined(geometry.clips[0], 'gamma render clip');
-      const middlePoint = expectDefined(
-        renderClip.points.find((point) => point.keyframeId === 'gamma-middle'),
-        'gamma middle point'
-      );
-      const middleCenterY = middlePoint.rect.y + middlePoint.rect.height / 2;
-
-      expect(middleCenterY).toBeCloseTo(56);
-      expect(renderClip.segments).toHaveLength(2);
-    });
-
-    it('sets, updates, and evaluates registered keyframes with side-aware interpolation', () => {
-      const addEvent = vi.fn();
-      const updateEvent = vi.fn();
-      engine.on('keyframe:add', addEvent);
-      engine.on('keyframe:update', updateEvent);
-
-      expect(
-        engine.setClipKeyframe({
-          clipId: 'clip1',
-          property: 'opacity',
-          time: fromSeconds(1),
-          value: 1,
-        })?.value
-      ).toBe(1);
-      engine.setClipKeyframe({
-        clipId: 'clip1',
-        property: 'opacity',
-        time: fromSeconds(3),
-        value: 0,
-        outgoing: { interpolation: 'hold' },
-      });
-      engine.setClipKeyframe({
-        clipId: 'clip1',
-        property: 'opacity',
-        time: fromSeconds(5),
-        value: 1,
-      });
-
-      expect(engine.getClipPropertyValueAtTime('clip1', 'opacity', fromSeconds(2))).toBe(0.5);
-      expect(engine.getClipPropertyValueAtTime('clip1', 'opacity', fromSeconds(4))).toBe(0);
-      expect(engine.getClipPropertyValueAtTime('clip1', 'opacity', fromSeconds(8))).toBeUndefined();
-
-      const middle = expectDefined(
-        engine.getClipKeyframes('clip1', 'opacity')[1],
-        'middle keyframe'
-      );
-      const updatedBySet = engine.setClipKeyframe({
-        clipId: 'clip1',
-        property: 'opacity',
-        time: fromSeconds(3),
-        value: 0.25,
-      });
-
-      expect(updatedBySet?.outgoing?.interpolation).toBe('hold');
-      expect(engine.getClipPropertyValueAtTime('clip1', 'opacity', fromSeconds(4))).toBe(0.25);
-
-      const updated = engine.updateClipKeyframe({
-        clipId: 'clip1',
-        keyframeId: middle.id,
-        value: 0.25,
-        outgoing: { interpolation: 'linear' },
-      });
-
-      expect(updated?.value).toBe(0.25);
-      expect(engine.getClipPropertyValueAtTime('clip1', 'opacity', fromSeconds(4))).toBe(0.625);
-
-      const bezier = engine.updateClipKeyframeSides({
-        clipId: 'clip1',
-        keyframeId: middle.id,
-        incoming: {
-          interpolation: 'bezier',
-          handle: { x: 0.8, y: 0 },
-        },
-        outgoing: {
-          interpolation: 'bezier',
-          handle: { x: 0.2, y: 1 },
-        },
-      });
-      const expectedBezierValue =
-        0.25 + (1 - 0.25) * getTimelineKeyframeBezierProgress(0.5, { x: 0.2, y: 1 }, undefined);
-
-      expect(bezier?.incoming).toEqual({ interpolation: 'bezier', handle: { x: 0.8, y: 0 } });
-      expect(bezier?.outgoing).toEqual({ interpolation: 'bezier', handle: { x: 0.2, y: 1 } });
-      expect(engine.getClipPropertyValueAtTime('clip1', 'opacity', fromSeconds(4))).toBeCloseTo(
-        expectedBezierValue
-      );
-
-      const defaultedHandle = engine.updateClipKeyframeSide({
-        clipId: 'clip1',
-        keyframeId: middle.id,
-        side: 'outgoing',
-        patch: { interpolation: 'bezier', handle: null },
-      });
-      expect(defaultedHandle?.outgoing).toEqual({
-        interpolation: 'bezier',
-        handle: { x: 0.42, y: 0 },
-      });
-
-      const reset = engine.updateClipKeyframeSide({
-        clipId: 'clip1',
-        keyframeId: middle.id,
-        side: 'outgoing',
-        patch: { interpolation: 'linear' },
-      });
-      expect(reset?.outgoing?.interpolation).toBe('linear');
-
-      expect(addEvent).toHaveBeenCalledTimes(3);
-      expect(updateEvent).toHaveBeenCalledTimes(5);
-    });
-
-    it('does not inherit side interpolation from neighboring keyframes', () => {
-      const first = engine.setClipKeyframe({
-        clipId: 'clip1',
-        property: 'opacity',
-        time: fromSeconds(4),
-        value: 1,
-      });
-      expect(first?.incoming).toBeUndefined();
-      expect(first?.outgoing).toBeUndefined();
-
-      engine.setClipKeyframe({
-        clipId: 'clip1',
-        property: 'opacity',
-        time: fromSeconds(1),
-        value: 0,
-        outgoing: { interpolation: 'bezier', handle: { x: 0.2, y: 0.8 } },
-      });
-
-      const independent = engine.setClipKeyframe({
-        clipId: 'clip1',
-        property: 'opacity',
-        time: fromSeconds(2),
-        value: 0.5,
-      });
-      expect(independent?.incoming).toBeUndefined();
-      expect(independent?.outgoing).toBeUndefined();
-
-      const explicit = engine.setClipKeyframe({
-        clipId: 'clip1',
-        property: 'opacity',
-        time: fromSeconds(3),
-        value: 0.75,
-        incoming: { interpolation: 'linear' },
-      });
-      expect(explicit?.incoming?.interpolation).toBe('linear');
-      expect(explicit?.outgoing).toBeUndefined();
-    });
-
-    it('rejects unregistered keyframe properties', () => {
-      expect(
-        engine.setClipKeyframe({
-          clipId: 'clip1',
-          property: 'volume',
-          time: fromSeconds(2),
-          value: 0.5,
-        })
-      ).toBeNull();
-      expect(engine.getClipPropertyValueAtTime('clip1', 'volume', fromSeconds(2))).toBeUndefined();
-    });
-
-    it('keeps neighboring keyframes when preview updates collide', () => {
-      const track = createKeyframeTrack([createKeyframeClip('kf-clip', 0, 10, [2, 3])]);
-      const previewEngine = new TimelineEngine({
-        tracks: [track],
-        keyframeProperties: [opacityKeyframeProperty],
-      });
-      const dragged = expectDefined(
-        previewEngine.getClipKeyframes('kf-clip')[1],
-        'dragged keyframe'
-      );
-
-      const preview = previewEngine.updateClipKeyframe(
-        { clipId: 'kf-clip', keyframeId: dragged.id, time: fromSeconds(2) },
-        { commit: false }
-      );
-
-      expect(toSeconds(expectDefined(preview, 'preview keyframe').time)).toBe(3);
-      expect(keyframeSeconds(previewEngine.getClip('kf-clip')?.clip)).toEqual([2, 3]);
-
-      const committed = previewEngine.updateClipKeyframe({
-        clipId: 'kf-clip',
-        keyframeId: dragged.id,
-        time: fromSeconds(2),
-      });
-
-      expect(toSeconds(expectDefined(committed, 'committed keyframe').time)).toBe(2);
-      expect(keyframeSeconds(previewEngine.getClip('kf-clip')?.clip)).toEqual([2]);
-    });
-
-    it('clones side interpolation state without property-specific fallbacks', () => {
-      const sideEngine = new TimelineEngine({
-        tracks: [
-          {
-            ...mockTrack,
-            clips: [
-              {
-                ...mockClip,
-                keyframes: [
-                  {
-                    id: 'linear-with-handle',
-                    property: 'opacity',
-                    time: fromSeconds(2),
-                    value: 0.5,
-                    outgoing: { interpolation: 'linear', handle: { x: 0.2, y: 0.8 } },
-                  },
-                  {
-                    id: 'bezier-without-handle',
-                    property: 'opacity',
-                    time: fromSeconds(3),
-                    value: 0.75,
-                    incoming: { interpolation: 'bezier' },
-                  },
-                ],
-              },
-            ],
-          },
-        ],
-        keyframeProperties: [opacityKeyframeProperty],
-      });
-
-      expect(sideEngine.getClipKeyframes('clip1')[0].outgoing).toEqual({
-        interpolation: 'linear',
-      });
-      expect(sideEngine.getClipKeyframes('clip1')[1].incoming).toEqual({
-        interpolation: 'bezier',
-        handle: { x: 0.58, y: 1 },
-      });
-    });
-
-    it('exposes selected keyframe geometry and hit testing', () => {
-      const keyframe = engine.setClipKeyframe({
-        clipId: 'clip1',
-        property: 'opacity',
-        time: fromSeconds(3),
-        value: 0.5,
-      });
-      const selectedKeyframe = expectDefined(keyframe, 'selected keyframe');
-
-      engine.selectClip('clip1');
-      engine.selectClipKeyframe('clip1', selectedKeyframe.id);
-      const rects = engine.getKeyframeRects({
-        selectedClipOnly: true,
-        rulerHeight: 32,
-        trackHeight: 48,
-        keyframeSize: 12,
-      });
-
-      expect(rects).toHaveLength(1);
-      expect(rects[0].keyframe.selected).toBe(true);
-      const hit = engine.getKeyframeAtPoint({
-        x: rects[0].rect.x + rects[0].rect.width / 2,
-        y: rects[0].rect.y + rects[0].rect.height / 2,
-        rulerHeight: 32,
-        trackHeight: 48,
-        keyframeSize: 12,
-      });
-
-      expect(hit?.keyframe.id).toBe(selectedKeyframe.id);
-    });
-
-    it('keeps edge keyframe hit rects inside clip bounds', () => {
-      const edgeEngine = new TimelineEngine({
-        tracks: [
-          {
-            ...mockTrack,
-            clips: [
-              {
-                ...mockClip,
-                keyframes: [
-                  {
-                    id: 'opacity-start',
-                    property: 'opacity',
-                    time: fromSeconds(1),
-                    value: 1,
-                  },
-                  {
-                    id: 'opacity-end',
-                    property: 'opacity',
-                    time: fromSeconds(5),
-                    value: 0,
-                  },
-                ],
-              },
-            ],
-          },
-        ],
-        zoomScale: 100,
-        keyframeProperties: [opacityKeyframeProperty],
-      });
-
-      const clipRect = expectDefined(
-        edgeEngine.getClipRect('clip1', { rulerHeight: 32, trackHeight: 48 }),
-        'clip rect'
-      );
-      const rects = edgeEngine.getKeyframeRects({
-        rulerHeight: 32,
-        trackHeight: 48,
-        keyframeSize: 12,
-      });
-
-      expect(rects).toHaveLength(2);
-      expect(rects[0].rect.x).toBe(clipRect.x);
-      expect(rects[1].rect.x + rects[1].rect.width).toBe(clipRect.x + clipRect.width);
-
-      const startHit = edgeEngine.getKeyframeAtPoint({
-        x: clipRect.x + 1,
-        y: rects[0].rect.y + rects[0].rect.height / 2,
-        rulerHeight: 32,
-        trackHeight: 48,
-        keyframeSize: 12,
-      });
-      const endHit = edgeEngine.getKeyframeAtPoint({
-        x: clipRect.x + clipRect.width - 1,
-        y: rects[1].rect.y + rects[1].rect.height / 2,
-        rulerHeight: 32,
-        trackHeight: 48,
-        keyframeSize: 12,
-      });
-
-      expect(startHit?.keyframe.id).toBe('opacity-start');
-      expect(endHit?.keyframe.id).toBe('opacity-end');
-    });
-
-    it('exposes keyframe segments and Bezier tangent handles with shared control point math', () => {
-      const segmentEngine = new TimelineEngine({
-        tracks: [
-          {
-            ...mockTrack,
-            clips: [
-              {
-                ...mockClip,
-                keyframes: [
-                  {
-                    id: 'opacity-start',
-                    property: 'opacity',
-                    time: fromSeconds(1),
-                    value: 0.2,
-                    outgoing: { interpolation: 'bezier', handle: { x: 0.2, y: 0.8 } },
-                    selected: true,
-                  },
-                  {
-                    id: 'opacity-middle',
-                    property: 'opacity',
-                    time: fromSeconds(3),
-                    value: 0.8,
-                    incoming: { interpolation: 'bezier', handle: { x: 0.8, y: 0.2 } },
-                    outgoing: { interpolation: 'hold' },
-                  },
-                  {
-                    id: 'opacity-end',
-                    property: 'opacity',
-                    time: fromSeconds(5),
-                    value: 0.4,
-                  },
-                ],
-              },
-            ],
-          },
-        ],
-        zoomScale: 100,
-        keyframeProperties: [opacityKeyframeProperty],
-      });
-
-      const segments = segmentEngine.getKeyframeSegments({
-        property: 'opacity',
-        rulerHeight: 32,
-        trackHeight: 48,
-        keyframeSize: 6,
-        tangentHandleSize: 8,
-        keyframeValuePadding: 7,
-      });
-
-      expect(segments).toHaveLength(2);
-      expect(segments[0].interpolation).toBe('bezier');
-      expect(segments[0].handles).toHaveLength(2);
-      expect(segments[1].interpolation).toBe('hold');
-      expect(segments[1].handles).toHaveLength(0);
-
-      const expectedControlPoints = getTimelineKeyframeBezierControlPoints(
-        segments[0].startPoint,
-        segments[0].endPoint,
-        { x: 0.2, y: 0.8 },
-        { x: 0.8, y: 0.2 }
-      );
-
-      expect(segments[0].controlPoint1).toEqual(expectedControlPoints.controlPoint1);
-      expect(segments[0].controlPoint2).toEqual(expectedControlPoints.controlPoint2);
-      expect(segments[0].handles[0].point).toEqual(expectedControlPoints.controlPoint1);
-      expect(segments[0].handles[1].point).toEqual(expectedControlPoints.controlPoint2);
-    });
-
-    it('hit-tests Bezier tangent handles with pointer padding', () => {
-      const segmentEngine = new TimelineEngine({
-        tracks: [
-          {
-            ...mockTrack,
-            clips: [
-              {
-                ...mockClip,
-                keyframes: [
-                  {
-                    id: 'opacity-start',
-                    property: 'opacity',
-                    time: fromSeconds(1),
-                    value: 0.2,
-                    outgoing: { interpolation: 'bezier', handle: { x: 0.25, y: 0.75 } },
-                  },
-                  {
-                    id: 'opacity-end',
-                    property: 'opacity',
-                    time: fromSeconds(5),
-                    value: 0.8,
-                    incoming: { interpolation: 'bezier', handle: { x: 0.75, y: 0.25 } },
-                  },
-                ],
-              },
-            ],
-          },
-        ],
-        zoomScale: 100,
-        keyframeProperties: [opacityKeyframeProperty],
-      });
-
-      const segment = expectDefined(
-        segmentEngine.getKeyframeSegments({
-          property: 'opacity',
-          rulerHeight: 32,
-          trackHeight: 48,
-          tangentHandleSize: 8,
-        })[0],
-        'keyframe segment'
-      );
-      const outgoing = expectDefined(segment.handles[0], 'outgoing handle');
-      const exactHit = segmentEngine.getKeyframeTangentHandleAtPoint({
-        property: 'opacity',
-        x: outgoing.point.x,
-        y: outgoing.point.y,
-        rulerHeight: 32,
-        trackHeight: 48,
-        tangentHandleSize: 8,
-      });
-      const touchHit = segmentEngine.getKeyframeTangentHandleAtPoint({
-        property: 'opacity',
-        x: outgoing.rect.x - 6,
-        y: outgoing.rect.y - 6,
-        pointerType: 'touch',
-        rulerHeight: 32,
-        trackHeight: 48,
-        tangentHandleSize: 8,
-      });
-
-      expect(exactHit?.side).toBe('outgoing');
-      expect(touchHit?.side).toBe('outgoing');
-    });
-
-    it('filters keyframe segments by selected keyframes and viewport visibility', () => {
-      const segmentEngine = new TimelineEngine({
-        tracks: [
-          createKeyframeTrack([
-            {
-              ...createKeyframeClip('curve-filter', 0, 10, [-1, 2, 8, 12]),
-              keyframes: [
-                {
-                  id: 'out-of-range-start',
-                  property: 'opacity',
-                  time: fromSeconds(-1),
-                  value: 0.1,
-                  outgoing: { interpolation: 'bezier' },
-                },
-                {
-                  id: 'visible-start',
-                  property: 'opacity',
-                  time: fromSeconds(2),
-                  value: 0.3,
-                  outgoing: { interpolation: 'bezier', handle: { x: 0.2, y: 1 } },
-                  selected: true,
-                },
-                {
-                  id: 'visible-end',
-                  property: 'opacity',
-                  time: fromSeconds(8),
-                  value: 0.7,
-                  incoming: { interpolation: 'bezier', handle: { x: 0.8, y: 0 } },
-                },
-                {
-                  id: 'out-of-range-end',
-                  property: 'opacity',
-                  time: fromSeconds(12),
-                  value: 1,
-                },
-              ],
-            },
-          ]),
-        ],
-        zoomScale: 100,
-        keyframeProperties: [opacityKeyframeProperty],
-      });
-
-      const selectedSegments = segmentEngine.getKeyframeSegments({
-        property: 'opacity',
-        selectedKeyframeOnly: true,
-        rulerHeight: 32,
-        trackHeight: 48,
-      });
-      const visibleSegments = segmentEngine.getVisibleKeyframeSegments({
-        property: 'opacity',
-        viewportWidth: 150,
-        rulerHeight: 32,
-        trackHeight: 48,
-      });
-
-      expect(selectedSegments.map((segment) => segment.segmentId)).toEqual([
-        'curve-filter:visible-start:visible-end:opacity',
-      ]);
-      expect(visibleSegments).toHaveLength(0);
-    });
-
-    it('keeps keyframes aligned when clips move and trims them across splits', () => {
-      engine.setClipKeyframe({
-        clipId: 'clip1',
-        property: 'opacity',
-        time: fromSeconds(2),
-        value: 0.2,
-        outgoing: { interpolation: 'bezier', handle: { x: 0.2, y: 0.8 } },
-      });
-      engine.setClipKeyframe({
-        clipId: 'clip1',
-        property: 'opacity',
-        time: fromSeconds(4),
-        value: 0.8,
-        incoming: { interpolation: 'bezier', handle: { x: 0.8, y: 0.2 } },
-      });
-
-      engine.moveClip({ clipId: 'clip1', startTime: fromSeconds(3) });
-      const movedKeyframes = engine.getClipKeyframes('clip1');
-      expect(movedKeyframes.map((keyframe) => toSeconds(keyframe.time))).toEqual([4, 6]);
-      expect(movedKeyframes[0].outgoing).toEqual({
-        interpolation: 'bezier',
-        handle: { x: 0.2, y: 0.8 },
-      });
-      expect(movedKeyframes[1].incoming).toEqual({
-        interpolation: 'bezier',
-        handle: { x: 0.8, y: 0.2 },
-      });
-
-      expect(engine.splitClip('clip1', fromSeconds(5))).toBe(true);
-      const clips = engine.getState().tracks[0].clips;
-      expect(clips).toHaveLength(2);
-      expect(clips[0].keyframes?.map((keyframe) => toSeconds(keyframe.time))).toEqual([4]);
-      expect(clips[1].keyframes?.map((keyframe) => toSeconds(keyframe.time))).toEqual([6]);
-      expect(clips[0].keyframes?.[0].outgoing).toEqual({
-        interpolation: 'bezier',
-        handle: { x: 0.2, y: 0.8 },
-      });
-      expect(clips[1].keyframes?.[0].incoming).toEqual({
-        interpolation: 'bezier',
-        handle: { x: 0.8, y: 0.2 },
-      });
-    });
-
-    it('removes out-of-range keyframes when clips are trimmed directly', () => {
-      const trimEngine = new TimelineEngine({
-        tracks: [createKeyframeTrack([createKeyframeClip('trimmed', 0, 10, [1, 3, 5, 8])])],
-        keyframeProperties: [opacityKeyframeProperty],
-      });
-
-      trimEngine.trimClip('trimmed', 'start', fromSeconds(2));
-      expect(keyframeSeconds(trimEngine.getClip('trimmed')?.clip)).toEqual([3, 5, 8]);
-
-      trimEngine.trimClip('trimmed', 'end', fromSeconds(6));
-      expect(keyframeSeconds(trimEngine.getClip('trimmed')?.clip)).toEqual([3, 5]);
-    });
-
-    it('preserves selected clip keyframes through cut and shifts them on paste', () => {
-      const cutEngine = new TimelineEngine({
-        tracks: [createKeyframeTrack([createKeyframeClip('cut-source', 1, 5, [2, 4])])],
-        keyframeProperties: [opacityKeyframeProperty],
-      });
-
-      cutEngine.selectClip('cut-source');
-      cutEngine.cutSelection();
-      expect(cutEngine.tracks[0].clips).toHaveLength(0);
-
-      cutEngine.pasteSelection(fromSeconds(10), 'keyframe-track');
-
-      const pastedClip = expectDefined(cutEngine.tracks[0].clips[0], 'pasted clip');
-      expect(pastedClip.id).not.toBe('cut-source');
-      expect(toSeconds(pastedClip.timelineStart)).toBe(10);
-      expect(keyframeSeconds(pastedClip)).toEqual([11, 13]);
-    });
-
-    it('filters keyframes when range removals trim clip edges', () => {
-      const deleteRangeEngine = new TimelineEngine({
-        tracks: [createKeyframeTrack([createKeyframeClip('delete-victim', 0, 10, [1, 4, 8])])],
-        keyframeProperties: [opacityKeyframeProperty],
-      });
-
-      deleteRangeEngine.commitEdit({
-        type: 'delete-range',
-        startTime: fromSeconds(0),
-        endTime: fromSeconds(3),
-        trackIds: ['keyframe-track'],
-      });
-
-      const rippleTrimmed = expectDefined(
-        deleteRangeEngine.getClip('delete-victim')?.clip,
-        'ripple trimmed clip'
-      );
-      expect(toSeconds(rippleTrimmed.timelineStart)).toBe(0);
-      expect(toSeconds(rippleTrimmed.timelineEnd)).toBe(7);
-      expect(keyframeSeconds(rippleTrimmed)).toEqual([1, 5]);
-
-      const liftRangeEngine = new TimelineEngine({
-        tracks: [createKeyframeTrack([createKeyframeClip('lift-victim', 0, 10, [2, 6, 8])])],
-        keyframeProperties: [opacityKeyframeProperty],
-      });
-
-      liftRangeEngine.commitEdit({
-        type: 'lift-range',
-        startTime: fromSeconds(7),
-        endTime: fromSeconds(12),
-        trackIds: ['keyframe-track'],
-      });
-
-      const endTrimmed = expectDefined(
-        liftRangeEngine.getClip('lift-victim')?.clip,
-        'end trimmed clip'
-      );
-      expect(toSeconds(endTrimmed.timelineStart)).toBe(0);
-      expect(toSeconds(endTrimmed.timelineEnd)).toBe(7);
-      expect(keyframeSeconds(endTrimmed)).toEqual([2, 6]);
-    });
-
-    it('filters keyframes when overwrite edits trim clip edges', () => {
-      const overwriteStartEngine = new TimelineEngine({
-        tracks: [createKeyframeTrack([createKeyframeClip('start-victim', 0, 10, [1, 4, 8])])],
-        keyframeProperties: [opacityKeyframeProperty],
-      });
-
-      overwriteStartEngine.commitEdit({
-        type: 'overwrite',
-        clip: createKeyframeClip('start-winner', 0, 3, []),
-        targetTrackId: 'keyframe-track',
-        startTime: fromSeconds(0),
-        snap: false,
-      });
-
-      const startTrimmed = expectDefined(
-        overwriteStartEngine.getClip('start-victim')?.clip,
-        'start trimmed victim'
-      );
-      expect(toSeconds(startTrimmed.timelineStart)).toBe(3);
-      expect(keyframeSeconds(startTrimmed)).toEqual([4, 8]);
-
-      const overwriteEndEngine = new TimelineEngine({
-        tracks: [createKeyframeTrack([createKeyframeClip('end-victim', 0, 10, [1, 4, 8])])],
-        keyframeProperties: [opacityKeyframeProperty],
-      });
-
-      overwriteEndEngine.commitEdit({
-        type: 'overwrite',
-        clip: createKeyframeClip('end-winner', 0, 3, []),
-        targetTrackId: 'keyframe-track',
-        startTime: fromSeconds(7),
-        snap: false,
-      });
-
-      const endTrimmed = expectDefined(
-        overwriteEndEngine.getClip('end-victim')?.clip,
-        'end trimmed victim'
-      );
-      expect(toSeconds(endTrimmed.timelineEnd)).toBe(7);
-      expect(keyframeSeconds(endTrimmed)).toEqual([1, 4]);
     });
   });
